A2C: fixed discounted reward calculation. value loss now in acceptable range
Using OpenAIs reward discounting makes sense mathematically. the value
loss is now not in the 200+ range anymore.
BUT learning is not happening anymore (for the most part).
At this point, so noise is added and one one thread is used. This will
most likely be one core issue right now.